Tests

+++ src/types.rs
@@ -1,4 +1,8 @@
+use rust_decimal::Decimal;
+
+// todo: it might be safer to use newtype pattern
+// but it adds a lot of boilerplate here so I decided not to
+// as long as those are different types and cannot be mixed-up that is
 pub type TransactionId = u32;
 pub type AccountId = u16;
-// todo: decimal
-pub type Currency = f32;
+pub type Currency = Decimal;
